requirements.txt
Django==4.2.5

# Api rest
djangorestframework==3.14.0

# Settings
django-environ==0.11.2
